move helper method to where it is only used (in lite impl)
